requirements.txt
# Core Cryptography
cryptography>=41.0.0
pycryptodome>=3.19.0
PyNaCl>=1.5.0

# Password Hashing
argon2-cffi>=23.1.0

# Authentication & TOTP
pyotp>=2.9.0
qrcode>=7.4.2
Pillow>=10.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality
black>=23.0.0
flake8>=6.0.0
pylint>=3.0.0

# Utilities
requests>=2.31.0

# UI (optional)
streamlit>=1.40.0
